ledger-on-sql: Allow `Long` values as the heartbeat timestamp.
`INTEGER` really does mean 32-bit, apparently.

+++ ledger/ledger-on-sql/src/main/resources/com/daml/ledger/on/sql/migrations/h2/V3__add_heartbeats_to_log.sql
@@ -8,7 +8,7 @@ ALTER TABLE ${table.prefix}log
     ALTER COLUMN envelope SET NULL;
 
 ALTER TABLE ${table.prefix}log
-    ADD COLUMN heartbeat_timestamp INTEGER;
+    ADD COLUMN heartbeat_timestamp BIGINT;
 
 ALTER TABLE ${table.prefix}log
     ADD CONSTRAINT record_or_timestamp CHECK (
