+++ ledger/ledger-on-sql/src/main/resources/com/daml/ledger/on/sql/migrations/postgresql/V3__add_heartbeats_to_log.sql
@@ -4,7 +4,7 @@
 ALTER TABLE ${table.prefix}log
     ALTER COLUMN entry_id DROP NOT NULL,
     ALTER COLUMN envelope DROP NOT NULL,
-    ADD COLUMN heartbeat_timestamp INTEGER,
+    ADD COLUMN heartbeat_timestamp BIGINT,
     ADD CONSTRAINT record_or_timestamp CHECK (
             (entry_id IS NOT NULL AND envelope IS NOT NULL AND heartbeat_timestamp IS NULL)
             OR (entry_id IS NULL AND envelope IS NULL AND heartbeat_timestamp IS NOT NULL)
